Se implementa la validacion para la documentacion y la funcionalidad para la actualizacion del estatus del derechohabiente para inicir afiliacion

+++ src/main/java/issstep/afiliacion/service/ArchivoService.java
@@ -15,15 +15,20 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import issstep.afiliacion.db.ArchivoDB;
 import issstep.afiliacion.db.UsuarioDB;
 import issstep.afiliacion.db.BeneficiarioDB;
+import issstep.afiliacion.db.DerechohabienteDB;
 import issstep.afiliacion.model.Archivo;
 import issstep.afiliacion.model.Mensaje;
+import issstep.afiliacion.model.ResultadoValidacion;
+import issstep.afiliacion.model.Usuario;
 import issstep.afiliacion.model.Beneficiario;
+import issstep.afiliacion.model.Derechohabiente;
 import issstep.afiliacion.utils.UtilsImage;
 
 import javax.servlet.http.*;
@@ -49,10 +54,24 @@ public class ArchivoService{
    
    @Autowired
    BeneficiarioDB beneficiarioDB;
+   
+   @Autowired
+   DerechohabienteDB derechohabienteDB;
+   
   
 	public ResponseEntity<?> uploadDocumento( long noControl, long noPreAfiliacion, long claveParentesco, long claveTipoArchivo,  MultipartFile uploadedFile, HttpServletResponse response) {
+		
+		ResultadoValidacion resultadoValidacion = validaEstatusRegistro (noControl, noPreAfiliacion);
+		
+		if (!resultadoValidacion.isEtatus())
+			return new ResponseEntity<>(new Mensaje(resultadoValidacion.getMensaje()), HttpStatus.CONFLICT);
+		
 		try{
-						
+			Archivo archivo = archivoDB.getArchivoByPersonaAndParentesco(noControl, noPreAfiliacion, claveParentesco,  claveTipoArchivo);
+			
+			if (archivo != null)
+				return new ResponseEntity<>(new Mensaje("Ya existe un archivo"), HttpStatus.CONFLICT);
+			
 			long idArchivoRegistrado;
 			
 			String desTipoDocto = archivoDB.getTipoArchivoByParentesco(claveParentesco, claveTipoArchivo);
@@ -70,8 +89,7 @@ public class ArchivoService{
 					
 					String resultado = UtilsImage.uploadFileToServer(UtilsImage.toPrettyURL(desTipoDocto), (MultipartFile) uploadedFile, "Persona"+noControl+noPreAfiliacion+claveParentesco+claveTipoArchivo+"-"+UtilsImage.toPrettyURL(desTipoDocto));
 					
-					
-					Archivo archivo = new Archivo();
+					archivo = new Archivo();
 					
 					archivo.setNoControl(noControl);
 					archivo.setNoPreAfiliacion(noPreAfiliacion);
@@ -111,32 +129,35 @@ public class ArchivoService{
 	}
 	
 	public ResponseEntity<?> updateDocumento( long claveDocumento,  MultipartFile uploadedFile, HttpServletResponse response) {
-		try{
+		try{	
+			Archivo archivo = archivoDB.getArchivo(claveDocumento);
 			
+			ResultadoValidacion resultadoValidacion = validaEstatusRegistro (archivo.getNoControl(), archivo.getNoPreAfiliacion());
 			
-				Archivo archivo = archivoDB.getArchivo(claveDocumento);
-				
-				if (archivo == null) 
-					return new ResponseEntity<>(new Mensaje("El archivo no existe"), HttpStatus.NO_CONTENT);
-		
-				// System.out.println("Eliminacion de informacion");
-				UtilsImage.deleteDocto(archivo.getUrlArchivo());
-				
-				String desTipoDocto = archivoDB.getTipoArchivoByParentesco(archivo.getClaveParentesco(), archivo.getClaveTipoArchivo());
-				
-				String resultado = UtilsImage.uploadFileToServer(UtilsImage.toPrettyURL(desTipoDocto), 
-																 (MultipartFile) uploadedFile, 
-																 "Persona"+archivo.getNoControl()+archivo.getNoPreAfiliacion()+archivo.getClaveParentesco()+archivo.getClaveTipoArchivo()+"-"+UtilsImage.toPrettyURL(desTipoDocto));
-				archivo.setNombre(uploadedFile.getOriginalFilename());
-				archivo.setUrlArchivo(resultado);
-				archivo.setEsValido(2);
-				archivo.setEstatus(1);
-				archivo.setFechaRegistro(new Timestamp(new Date().getTime()));	
-				
-				if (archivoDB.update(archivo) == -1)
-					return new ResponseEntity<>(new Mensaje("No fue posible elimiar el registro del archivo"), HttpStatus.INTERNAL_SERVER_ERROR);
-				
-				return new ResponseEntity<>(new Mensaje("Archivo actualizado"), HttpStatus.OK);
+			if (!resultadoValidacion.isEtatus())
+				return new ResponseEntity<>(new Mensaje(resultadoValidacion.getMensaje()), HttpStatus.CONFLICT);
+			
+			if (archivo == null) 
+				return new ResponseEntity<>(new Mensaje("El archivo no existe"), HttpStatus.NO_CONTENT);
+	
+			// System.out.println("Eliminacion de informacion");
+			UtilsImage.deleteDocto(archivo.getUrlArchivo());
+			
+			String desTipoDocto = archivoDB.getTipoArchivoByParentesco(archivo.getClaveParentesco(), archivo.getClaveTipoArchivo());
+			
+			String resultado = UtilsImage.uploadFileToServer(UtilsImage.toPrettyURL(desTipoDocto), 
+															 (MultipartFile) uploadedFile, 
+															 "Persona"+archivo.getNoControl()+archivo.getNoPreAfiliacion()+archivo.getClaveParentesco()+archivo.getClaveTipoArchivo()+"-"+UtilsImage.toPrettyURL(desTipoDocto));
+			archivo.setNombre(uploadedFile.getOriginalFilename());
+			archivo.setUrlArchivo(resultado);
+			archivo.setEsValido(2);
+			archivo.setEstatus(1);
+			archivo.setFechaRegistro(new Timestamp(new Date().getTime()));	
+			
+			if (archivoDB.update(archivo) == -1)
+				return new ResponseEntity<>(new Mensaje("No fue posible elimiar el registro del archivo"), HttpStatus.INTERNAL_SERVER_ERROR);
+			
+			return new ResponseEntity<>(new Mensaje("Archivo actualizado"), HttpStatus.OK);
 		}
 		catch(Exception ex){
 			System.err.println("Exception ArchivoService.uploadDocto ");
@@ -153,6 +174,11 @@ public class ArchivoService{
 			// System.out.println(archivo);
 			
 			if(archivo != null) {
+				ResultadoValidacion resultadoValidacion = validaEstatusRegistro (archivo.getNoControl(), archivo.getNoPreAfiliacion());
+				
+				if (!resultadoValidacion.isEtatus())
+					return new ResponseEntity<>(new Mensaje(resultadoValidacion.getMensaje()), HttpStatus.CONFLICT);
+				
 			
 				String nombreTemp = archivo.getUrlArchivo();
 				InputStream imputStream = new FileInputStream(new File(nombreTemp));
@@ -217,11 +243,25 @@ public class ArchivoService{
 	}
 	
 	public ResponseEntity<?> deleteDocumento(long claveDocumento, HttpServletResponse response) {
+		
+		
 		try {
 			
 			Archivo archivo = archivoDB.getArchivo(claveDocumento);
 			if(archivo == null)
 				return new ResponseEntity<>(new Mensaje("Documento no encontrado"), HttpStatus.NOT_FOUND);
+			
+			ResultadoValidacion resultadoValidacion = validaEstatusRegistro (archivo.getNoControl(), archivo.getNoPreAfiliacion());
+			
+			if (!resultadoValidacion.isEtatus())
+				return new ResponseEntity<>(new Mensaje(resultadoValidacion.getMensaje()), HttpStatus.CONFLICT);
+			
+			Usuario usuarioLogin = getInfoLogin();
+			if (usuarioLogin == null)
+				return new ResponseEntity<>(new Mensaje("Usuario no autentificado"), HttpStatus.BAD_REQUEST);
+			
+			if (archivo.getEsValido() == 1 && !usuarioLogin.getRol().equals("ADMINISTRADOR")) 
+				return new ResponseEntity<>(new Mensaje("El estatus del documento es valido"), HttpStatus.NOT_FOUND);
 			
 			UtilsImage.deleteDocto(archivo.getUrlArchivo());
 			if (archivoDB.delete(claveDocumento) == -1)
@@ -237,7 +277,7 @@ public class ArchivoService{
 	}
 	
 	public ResponseEntity<?> updateValidacionDocto( long claveDocumento, int estatusValidacion, HttpServletResponse response) {
-		if (estatusValidacion != 0 && estatusValidacion != 1)
+		if (estatusValidacion < 0 || estatusValidacion > 2)
 			return new ResponseEntity<>(new Mensaje("Estatus invalido"), HttpStatus.BAD_REQUEST);
 		
 		try {
@@ -261,10 +301,38 @@ public class ArchivoService{
 		}
 	}
 
-
+	ResultadoValidacion validaEstatusRegistro(long noControl, long noPreAfiliacion) {
+		ResultadoValidacion resultadoValidacion = new ResultadoValidacion();
+		Derechohabiente derechohabiente =  derechohabienteDB.getPersonaByNoControlNoPreafiliacion(noControl, noPreAfiliacion);
+		
+		if (derechohabiente == null) {
+			resultadoValidacion.setEtatus(false);
+			resultadoValidacion.setMensaje("No existe el derechohabiente con número de control: " + noControl
+										+  "y número de afiliación: " + noPreAfiliacion);
+			return resultadoValidacion;
+		}
+		
+		int estatus = derechohabiente.getEstatus();
+		
+		if (estatus == 9)  {
+			resultadoValidacion.setEtatus(false);
+			resultadoValidacion.setMensaje("El derechohabiente no se puede actualizar");
+			return resultadoValidacion;
+		}
+		
+		resultadoValidacion.setEtatus(true);
+		resultadoValidacion.setMensaje("");
+		return resultadoValidacion;
+		
+	}
 	
+	Usuario getInfoLogin() {
+		String user = (String) SecurityContextHolder.getContext().getAuthentication().getName();
+		if (user == "anonymousUser") 
+			return null;			
+		Usuario usuario =  usuarioDB.getUsuarioByColumnaStringValor("LOGIN", user);
 	
-	
-	
+		return usuario;
+	}
 		
 }
